requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
psycopg2-binary
bcrypt==4.1.2
email-validator==2.1.0
numpy
joblib
flask_login
scikit-learn==1.6.1